requirements.txt
tensorflow
pandas
numpy
scikit-learn
streamlit
astroquery
plotly
lightkurve
google-generativeai
python-dotenv